Backend/requirements/base.txt
# Park & Go V2 — Base Dependencies (Python 3.13+)
# Install: pip install -r requirements/base.txt

# Framework
fastapi==0.115.6
starlette==0.41.3

# ASGI server
uvicorn[standard]==0.34.0

# Database & ORM (PostgreSQL + PostGIS)
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
GeoAlchemy2==0.14.3

# Migrations
alembic==1.14.1

# Config & validation
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# HTTP clients
httpx==0.28.1
requests==2.32.3

# Auth: JWT, password hashing, OAuth2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
oauthlib==3.2.2
cryptography==44.0.0

# Caching (Redis)
redis>=5.0.0

# Task queue (Celery)
celery[redis]==5.4.0

# Observability / metrics
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2

# Rate limiting
slowapi==0.1.9

# Campus event iCal
icalendar>=7.0.0

# Scheduling
apscheduler>=3.10.0
